backend/requirements.txt
chromadb
sentence-transformers
langchain-core
geopandas
shapely
langchain
langchain-openai
langchain-community
langchain-huggingface
langgraph
flask
flask-cors
python-dotenv
psycopg2-binary
sqlalchemy
geoalchemy2



